Refactor common utilities to screens/Utils file
