+++ app/src/test/java/com/mr3y/ludi/ui/screens/onboarding/accessibility/OnboardingScreenAccessibilityTest.kt
@@ -19,14 +19,14 @@ import com.mr3y.ludi.ui.presenter.model.FavouriteGame
 import com.mr3y.ludi.ui.presenter.model.NewsDataSource
 import com.mr3y.ludi.ui.presenter.model.actualResource
 import com.mr3y.ludi.ui.screens.BaseRobolectricTest
+import com.mr3y.ludi.ui.screens.getString
+import com.mr3y.ludi.ui.screens.onNodeWithStateDescription
 import com.mr3y.ludi.ui.screens.onboarding.FakeAllGenres
 import com.mr3y.ludi.ui.screens.onboarding.FakeNewsDataSources
 import com.mr3y.ludi.ui.screens.onboarding.FakeOnboardingGames
 import com.mr3y.ludi.ui.screens.onboarding.FakeOnboardingState
 import com.mr3y.ludi.ui.screens.onboarding.OnboardingScreen
-import com.mr3y.ludi.ui.screens.onboarding.getString
 import com.mr3y.ludi.ui.screens.onboarding.onGenre
-import com.mr3y.ludi.ui.screens.onboarding.onNodeWithStateDescription
 import com.mr3y.ludi.ui.theme.LudiTheme
 import org.junit.Test
